Fix DN (BaseDN) parameter docstrings
In X.500 parlance, a _Distinguished Name_ (_DN_) is a fully-qualified
object path<sup>[1][1],[2][2]</sup>  For example, `CN=example.com,
O=IANA, C=US` is a DN.  Prior to this, however, the `DN` parameter did
not specify the fully-qualified DN, but instead only the path part,
minus the _Common Name_.  In the above example, that would be `O=IANA,
C=US`, were the common name is `CN=example.com`.  This path part is
properly known as the _Base DN_.<sup>[2][2],[3][3]</sup>  Therefore,
this commit updates the documentation for the `DN` parameter with the
appropriate names.

Fixes #89

[1]: https://ldapwiki.com/wiki/Distinguished%20Names
[2]: https://www.novell.com/documentation/extend5/Docs/help/Composer/books/LDAPGlossary.html
[3]: https://ldapwiki.com/wiki/BaseDN