WSManListener: Only match server certs
Prior to this the `Find-Certificate` function was inconsistent about
checking for the `serverAuthentication` attribute, a requirement for
HTTPS servers such as WSMan listners.  This commit fixes this and takes
some inital steps to remove duplicate code that might lead to similar
oversights in the future.  To do this, it moves the
`serverAuthentication` filter from the `Where-Object` script block to a
certificate provider mixin for `Get-ChildItem`-- the
`-SSLServerAuthentication` parameter.  To reduce duplication of code,
it also moves the function's `Get-ChildItem` parameters, which are
always the same, into a private `PSDefaultParameters` block.