WSManListner/Find-Certificate: Rewrite as an expression compiler
Prior to this the `Find-Certificate` function in `WSManListner` used a
complex web of `if-then-else` statements to try successive combinatorial
conditions.  Adding additional features in that framework was daunting,
because the combinatorial explosion would require an expodential
addition of conditional statmenets.  Namely, I wanted to make it
possible to pass in multiple values for `Issuers` and/or `DN` to match
the variety of CA certificates I deal with.  Rather than adding more to
the pile, this commit refactors-- essentially rewrites-- the
`Find-Certificate` function to work more like a rudimentary compiler.
As a nice side-effect, I was able to add the multi-matcher feature.

The implementation starts with a `Hashtable` (`$X509Properites`) to
represent the filter criteria.  Each key in the hashtable represents a
potential property of a certificate, and the values represent possible
property values.  Once the `Hashtable` is set, this new implementation
compiles it into expressions, and joins them together using `-and`.  The
final expression string is then cast into a `ScriptBlock` and passed to
`Where-Object -FilterScript`.

A few tests are added to demonstrate the new multi-matcher feature.  The
mock call counters on all `Find-Certificate` tests have also been
updated, because `Get-ChildItem` is now only ever called once.

I've done my best to ensure this doesn't break any of the current tests
or documented functionality, but it is possible there are some edge
cases where this will break things.