Find-Certificate: Add test for multiple certs
Prior to this the tests for `Find-Certificate` only ever tested empty or
single-element CertStore mocks.  The documented behavior when
`SubjectFormat == Both` is to select the FQDN cert over the short name
cert, which requires a CertStore with two or more certificates.  This
commit modifies the mocks and adds a test for that condition.

This change endeavored not to change any of the existing tests, but due
to some changes in the mocks, a few minor changes were required.
Namely, the mock call counts for a few tests changed, because the
default certificate now uses the short hostname rather than the FQDN.
Absolutely non of the module code has changed, however, so the pure
functionality of the DSC resources remains intact.